+++ NEWS.md
@@ -4,9 +4,25 @@
 
 - Objective function and non-linear constraints for `cobyla`:
   - Non-linear equality constraints can be specified by keyword `nonlinear_eq`.
-  - The objective function is called as `f(x)` as for other algorithms.
+  - The objective function is called as `f(x)` like in other algorithms.
   - The functions implementing the non-linear constraints are passed by
     keywords `nonlinear_eq` and `nonlinear_ineq`.
+
+- Algorithms have a more similar interface:
+  - All algorithms have the same positional input, only the available keywords
+    may change.
+  - All algorithms have the same convention for the objective function.
+    Non-linear constraints, if any, are provided by other user-defined
+    functions.
+  - All algorithms yield a 2-tuple `(x,info)` with `x` an approximate solution
+    (provided algorithm was successful) and `info` a structured object
+    collecting other outputs from the algorithm. The following properties are
+    available for all algorithms: `info.fx` is the objective function value
+    `f(x)`, `info.nf` is the number of evaluations of the objective function,
+    and `info.status` is the termination status of the algorithm.
+  - `issuccess(info)` yields whether algorithm was successful.
+  - `PRIMA.reason(info)` yields a textual description of the termination status
+    of the algorithm.
 
 ## Version 0.1.1